Improve random island generation

+++ src/utils/consts.rs
@@ -1,1 +1,1 @@
-pub const TILE_SIZE: u8 = 32;
+pub const TILE_SIZE: u32 = 32;
